Add mig to remove tables and squash
The migrations are creating the tables but the models are gone. So when running makemigrations, a new one is created to delete the tables. This however introduces errors.
So I created the corresponding migration and created a squashedmigration to just avoid creating and deleting the same tables.
This however requires an update of the contribution_plan module that references one of these migrations directly.