OP-1557 Add dependency for contribution plan
The contribution plan migrations refer to the calculationrules model that is then deleted.
We need to ensure that these operations are executed in the right order, i.e. deleting the model only after the contribution plan doesn't need it anymore.

This shouldn't affect the development branch that has squashed commits and other operations. It also won't break the upgrade as the migrations themselves are the same, only the dependency tree.